+++ packages/main/src/FileUploader.ts
@@ -3,7 +3,6 @@ import customElement from "@ui5/webcomponents-base/dist/decorators/customElement
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
-import languageAware from "@ui5/webcomponents-base/dist/decorators/languageAware.js";
 import ValueState from "@ui5/webcomponents-base/dist/types/ValueState.js";
 import { getFeature } from "@ui5/webcomponents-base/dist/FeaturesRegistry.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
@@ -63,8 +62,20 @@ import type { IFormElement } from "./features/InputElementsFormSupport.js";
  * @tagname ui5-file-uploader
  * @public
  */
-@customElement("ui5-file-uploader")
-@languageAware
+@customElement({
+	tag: "ui5-file-uploader",
+	languageAware: true,
+	renderer: litRender,
+	styles: FileUploaderCss,
+	template: FileUploaderTemplate,
+	staticAreaTemplate: FileUploaderPopoverTemplate,
+	staticAreaStyles: [ResponsivePopoverCommonCss, ValueStateMessageCss],
+	dependencies: [
+		Input,
+		Popover,
+		Icon,
+	],
+})
 /**
  * Event is fired when the value of the file path has been changed.
  * <b>Note:</b> Keep in mind that because of the HTML input element of type file, the event is also fired in Chrome browser when the Cancel button of the uploads window is pressed.
@@ -238,26 +249,6 @@ class FileUploader extends UI5Element implements IFormElement {
 
 	static get formAssociated() {
 		return true;
-	}
-
-	static get render() {
-		return litRender;
-	}
-
-	static get styles() {
-		return FileUploaderCss;
-	}
-
-	static get template() {
-		return FileUploaderTemplate;
-	}
-
-	static get staticAreaTemplate() {
-		return FileUploaderPopoverTemplate;
-	}
-
-	static get staticAreaStyles() {
-		return [ResponsivePopoverCommonCss, ValueStateMessageCss];
 	}
 
 	constructor() {
@@ -508,10 +499,6 @@ class FileUploader extends UI5Element implements IFormElement {
 
 	get ui5Input() {
 		return this.shadowRoot!.querySelector<Input>(".ui5-file-uploader-input");
-	}
-
-	static get dependencies() {
-		return [Input, Popover, Icon];
 	}
 
 	static async onDefine() {