+++ packages/main/src/Tab.ts
@@ -1,6 +1,5 @@
 import UI5Element from "@ui5/webcomponents-base/dist/UI5Element.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
-import languageAware from "@ui5/webcomponents-base/dist/decorators/languageAware.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
 import type { ITabbable } from "@ui5/webcomponents-base/dist/delegate/ItemNavigation.js";
 import property from "@ui5/webcomponents-base/dist/decorators/property.js";
@@ -61,8 +60,18 @@ const DESIGN_DESCRIPTIONS = {
  * @implements sap.ui.webc.main.ITab
  * @public
  */
-@customElement("ui5-tab")
-@languageAware
+@customElement({
+	tag: "ui5-tab",
+	languageAware: true,
+	renderer: litRender,
+	template: TabTemplate,
+	styles: css,
+	dependencies: [
+		Icon,
+		Button,
+		CustomListItem,
+	],
+})
 class Tab extends UI5Element implements ITab, ITabbable {
 	/**
 	 * The text to be displayed for the item.
@@ -193,26 +202,6 @@ class Tab extends UI5Element implements ITab, ITabbable {
 	_getElementInStrip?: () => ITab | null;
 	_individualSlot!: string;
 
-	static get render() {
-		return litRender;
-	}
-
-	static get template() {
-		return TabTemplate;
-	}
-
-	static get stripTemplate() {
-		return TabInStripTemplate;
-	}
-
-	static get overflowTemplate() {
-		return TabInOverflowTemplate;
-	}
-
-	static get styles() {
-		return css;
-	}
-
 	static i18nBundle: I18nBundle;
 
 	set _tabIndex(val: string) {
@@ -221,14 +210,6 @@ class Tab extends UI5Element implements ITab, ITabbable {
 
 	get _tabIndex() {
 		return this.getTabInStripDomRef()!.getAttribute("tabindex")!;
-	}
-
-	static get dependencies() {
-		return [
-			Icon,
-			Button,
-			CustomListItem,
-		];
 	}
 
 	get displayText() {
@@ -288,6 +269,7 @@ class Tab extends UI5Element implements ITab, ITabbable {
 	 *
 	 * @function
 	 * @public
+     * @name sap.ui.webc.main.Tab.prototype.getTabInStripDomRef
 	 * @since 1.0.0-rc.16
 	 */
 	getTabInStripDomRef() {
@@ -464,6 +446,14 @@ class Tab extends UI5Element implements ITab, ITabbable {
 
 	get overflowState() {
 		return (this.disabled || this.isSingleClickArea) ? ListItemType.Inactive : ListItemType.Active;
+	}
+
+	static get stripTemplate() {
+		return TabInStripTemplate;
+	}
+
+	static get overflowTemplate() {
+		return TabInOverflowTemplate;
 	}
 
 	static async onDefine() {